Explicit integer conversion in read_plots_xml.

+++ src/input_xml.F90
@@ -1338,9 +1338,9 @@ contains
           pl % color_by = PLOT_COLOR_CELLS
           allocate(pl % colors(n_cells))
           do j = 1, n_cells
-            pl % colors(j) % rgb(1) = prn()*255
-            pl % colors(j) % rgb(2) = prn()*255
-            pl % colors(j) % rgb(3) = prn()*255
+            pl % colors(j) % rgb(1) = int(prn()*255)
+            pl % colors(j) % rgb(2) = int(prn()*255)
+            pl % colors(j) % rgb(3) = int(prn()*255)
           end do
 
         case ("mat", "material")
@@ -1348,9 +1348,9 @@ contains
           pl % color_by = PLOT_COLOR_MATS
           allocate(pl % colors(n_materials))
           do j = 1, n_materials
-            pl % colors(j) % rgb(1) = prn()*255
-            pl % colors(j) % rgb(2) = prn()*255
-            pl % colors(j) % rgb(3) = prn()*255
+            pl % colors(j) % rgb(1) = int(prn()*255)
+            pl % colors(j) % rgb(2) = int(prn()*255)
+            pl % colors(j) % rgb(3) = int(prn()*255)
           end do
 
         case default
